WIP: Handle terminal window resize
Catching SIGWINCH signals. Now we just need to tell the game to update
when the window is resized, which will include some new memory
allocations since we only allocate room for the snake to grow the area
of the window surface.

This will probably not work on Windows though due to the lack of
SIGWINCH, not sure what to do about that yet...